cec: do One Touch Play after boot [1/1]
PD#SWPL-50044

Problem:
new feature, customer would like the box to do
otp to wake up tv and tune to it after boot.

Solution:
do One Touch Play when the initiation mode is boot.

Verify:
newton

Change-Id: Ib479396ace7d9a9ae8bcd44d524963a1645e6df8
Signed-off-by: an.xi <an.xi@amlogic.com>

cec: Extend the overtime period for power query [1/1]

PD# SWPL-50368

Problem:
The audio method for judging whether the connected TV supports
cec is to send a message to the TV to query the power status,
and then wait for the message to be returned. The problem is
that the timeout period specified by the cec protocol is 2 seconds.
In fact, in the hotplug scene, TV may have a low probability of
replying to this message for more than 2 seconds, which will cause
a misjudgment.

Solution:
Extend the overtime period for power query. Add a method of checking
if the connected hdmi device supports cec by aosp could be done in
the future.

Verify:
Oppen

Change-Id: I9b46bdc4fe45fe3d7f7ac253c28de9a6bbbb6fd3
Signed-off-by: an.xi <an.xi@amlogic.com>