cec: update aml feature for dalton soundbar [1/4]
PD#SWPL-156042

Problem:
New feature of implenting soundbar functions on dalton.

Solution:
1.When connected with a tv, system audio control and arc
are established. The active source in the cec system is
the only condition to determine whether the soundbar enters
avr mode or not.
a. If the active source is outside of the soundbar, then
the device is in avr mode and working like an avr.
b. If the active source is the soundbar or a source device
connected to the soundbar, then the soundbar is in box mode
and working like a box.
2.As hdmi_arc_in device is the permanent audio device
on aml platform, don't add it to AudioService when
it's in avr mode but to control audio patch instead.

Verify:
dalton

Change-Id: I2a1afd89ce270f96bd7c5b31704ae655e0ae812c
Signed-off-by: an.xi <an.xi@amlogic.com>