SystemUI: Check whether the selected WFD route is available
When the selected WFD route is not available, the cast tile in
quick settings and the cast slot in the status bar are both shown
as a connected icon, it is not reasonable. This CL checks whether
the selected route is available when getting the cast devices.

Type: Fix
Redmine ID: #479239
Associated modifications: N/A
Test: N/A

Signed-off-by: jianhui.wang <wjh@rock-chips.com>
Change-Id: I3534901d09c01a9daaa241f97efca0d90cbe07b9